Calculate source-size length as a float https://bugs.webkit.org/show_bug.cgi?id=144766
Reviewed by Dean Jackson.

Source/WebCore:

Make sure that the source-size length is calculated as a float,
to align with the spec.

Test: fast/dom/HTMLImageElement/sizes/image-sizes-1x.html
      fast/dom/HTMLImageElement/sizes/image-sizes-2x.html

* css/SourceSizeList.cpp:
(WebCore::parseSizesAttribute):
(WebCore::defaultLength):
(WebCore::computeLength):
* css/SourceSizeList.h:
* html/HTMLImageElement.cpp:
(WebCore::HTMLImageElement::parseAttribute):
* html/parser/HTMLPreloadScanner.cpp:
(WebCore::TokenPreloadScanner::StartTagScanner::processAttributes):
* html/parser/HTMLSrcsetParser.cpp:
(WebCore::pickBestImageCandidate):
(WebCore::bestFitSourceForImageAttributes):
* html/parser/HTMLSrcsetParser.h:

LayoutTests:

Add tests that make sure fractional source-size lengths are supported
and are calculated as floats.

* fast/dom/HTMLImageElement/sizes/image-sizes-1x-expected.txt:
* fast/dom/HTMLImageElement/sizes/image-sizes-1x.html:
* fast/dom/HTMLImageElement/sizes/image-sizes-2x-expected.txt:
* fast/dom/HTMLImageElement/sizes/image-sizes-2x.html:


Canonical link: https://commits.webkit.org/162815@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183984 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/css/SourceSizeList.cpp
@@ -34,7 +34,7 @@ namespace WebCore {
 
 #if !ENABLE(PICTURE_SIZES)
 
-unsigned parseSizesAttribute(StringView, RenderView*, Frame*)
+float parseSizesAttribute(StringView, RenderView*, Frame*)
 {
     return 0;
 }
@@ -56,28 +56,28 @@ static bool match(std::unique_ptr<MediaQueryExp>&& expression, RenderStyle& styl
     return mediaQueryEvaluator.eval(mediaQuerySet.get());
 }
 
-static unsigned defaultLength(RenderStyle& style, RenderView* view)
+static float defaultLength(RenderStyle& style, RenderView* view)
 {
-    return CSSPrimitiveValue::create(100, CSSPrimitiveValue::CSS_VW)->computeLength<unsigned>(CSSToLengthConversionData(&style, &style, view));
+    return CSSPrimitiveValue::create(100, CSSPrimitiveValue::CSS_VW)->computeLength<float>(CSSToLengthConversionData(&style, &style, view));
 }
 
-static unsigned computeLength(CSSValue* value, RenderStyle& style, RenderView* view)
+static float computeLength(CSSValue* value, RenderStyle& style, RenderView* view)
 {
     CSSToLengthConversionData conversionData(&style, &style, view);
     if (is<CSSPrimitiveValue>(value)) {
         CSSPrimitiveValue& primitiveValue = downcast<CSSPrimitiveValue>(*value);
         if (!primitiveValue.isLength())
             return defaultLength(style, view);
-        return primitiveValue.computeLength<unsigned>(conversionData);
+        return primitiveValue.computeLength<float>(conversionData);
     }
     if (is<CSSCalcValue>(value)) {
         Length length(downcast<CSSCalcValue>(*value).createCalculationValue(conversionData));
-        return CSSPrimitiveValue::create(length, &style)->computeLength<unsigned>(conversionData);
+        return CSSPrimitiveValue::create(length, &style)->computeLength<float>(conversionData);
     }
     return defaultLength(style, view);
 }
 
-unsigned parseSizesAttribute(StringView sizesAttribute, RenderView* view, Frame* frame)
+float parseSizesAttribute(StringView sizesAttribute, RenderView* view, Frame* frame)
 {
     if (!view)
         return 0;
