+++ Source/WebCore/html/parser/HTMLPreloadScanner.cpp
@@ -111,7 +111,7 @@ public:
 
         // Resolve between src and srcSet if we have them and the tag is img.
         if (m_tagId == TagId::Img && !m_srcSetAttribute.isEmpty()) {
-            unsigned sourceSize = 0;
+            float sourceSize = 0;
 #if ENABLE(PICTURE_SIZES)
             sourceSize = parseSizesAttribute(m_sizesAttribute, document.renderView(), document.frame());
 #else
